refactor(downloader): share success-rate/statistics via DownloaderStatsMixin
GuideDownloader and SeriesDownloader carried identical copies of
_calculate_success_rate() and get_downloader_statistics() (series only
differing by an extra cache_efficiency key). Hoist both, plus a
_reset_counters() helper, into a DownloaderStatsMixin in downloader/base.py
that both now inherit; the stats dict always includes cache_efficiency.
